Move Column Management form into a separate component
